requirements.txt
tensorflow-gpu==2.0.0rc0
numpy
matplotlib
tqdm
